[master]: Change Error signature to take error directly.
This changes uhttp.Error to take the error object directly, and extract
the string from it via Error(). This should allow for more flexibility
to accept different kinds of errors in the future, and also more closely
matches the 90% use case.

Additionally, added a Status write method that constructs an appropriate
body from the status code alone, in order to simplify the common 204 &
403 use cases in the Utopia code.